+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 2.0.0
 
-- Migrated from PHP 5.5 to PHP 7.1
+- Migrated from PHP 5.5 to PHP 7.1+
 
 - Updated PHP HTTP dependency to version 2
 
@@ -17,3 +17,5 @@
 - Removed the `InternalException` only used in the `PhpArray` service
 
 - Modified the PHPArray service to only support scalars as rates (rate objects are not compatible)
+
+- Supporting both PSR-6 Cache and PSR-16 Simple Cache caching